+++ cmd/logcli/query.go
@@ -1,15 +1,12 @@
 package main
 
 import (
-	"fmt"
 	"log"
-	"sort"
 	"strings"
 	"time"
 
 	"github.com/fatih/color"
 	"github.com/prometheus/prometheus/pkg/labels"
-	"github.com/prometheus/prometheus/promql"
 
 	"github.com/grafana/loki/pkg/iter"
 	"github.com/grafana/loki/pkg/logproto"
@@ -52,15 +49,15 @@ func doQuery() {
 	}
 
 	if len(common) > 0 {
-		fmt.Println("Common labels:", color.RedString(common.String()))
+		log.Println("Common labels:", color.RedString(common.String()))
 	}
 
 	if len(*ignoreLabelsKey) > 0 {
-		fmt.Println("Ignoring labels key:", color.RedString(strings.Join(*ignoreLabelsKey, ",")))
+		log.Println("Ignoring labels key:", color.RedString(strings.Join(*ignoreLabelsKey, ",")))
 	}
 
 	// Get the max size of labels
-	maxLabelsLen := 0
+	maxLabelsLen := *fixedLabelsLen
 	for _, ls := range cache {
 		ls = subtract(common, ls)
 		if len(*ignoreLabelsKey) > 0 {
@@ -92,94 +89,4 @@ func doQuery() {
 	if err := i.Error(); err != nil {
 		log.Fatalf("Error from iterator: %v", err)
 	}
-}
-
-func printLogEntry(ts time.Time, lbls string, line string) {
-	fmt.Println(
-		color.BlueString(ts.Format(time.RFC3339)),
-		color.RedString(lbls),
-		strings.TrimSpace(line),
-	)
-}
-
-func padLabel(ls labels.Labels, maxLabelsLen int) string {
-	labels := ls.String()
-	if len(labels) < maxLabelsLen {
-		labels += strings.Repeat(" ", maxLabelsLen-len(labels))
-	}
-	return labels
-}
-
-func mustParseLabels(labels string) labels.Labels {
-	ls, err := promql.ParseMetric(labels)
-	if err != nil {
-		log.Fatalf("Failed to parse labels: %+v", err)
-	}
-	return ls
-}
-
-func parseLabels(resp *logproto.QueryResponse) (map[string]labels.Labels, []labels.Labels) {
-	cache := make(map[string]labels.Labels, len(resp.Streams))
-	lss := make([]labels.Labels, 0, len(resp.Streams))
-	for _, stream := range resp.Streams {
-		ls := mustParseLabels(stream.Labels)
-		cache[stream.Labels] = ls
-		lss = append(lss, ls)
-	}
-	return cache, lss
-}
-
-func commonLabels(lss []labels.Labels) labels.Labels {
-	if len(lss) == 0 {
-		return nil
-	}
-
-	result := lss[0]
-	for i := 1; i < len(lss); i++ {
-		result = intersect(result, lss[i])
-	}
-	return result
-}
-
-// intersect two labels set
-func intersect(a, b labels.Labels) labels.Labels {
-
-	set := labels.Labels{}
-	ma := a.Map()
-	mb := b.Map()
-
-	for ka, va := range ma {
-		if vb, ok := mb[ka]; ok {
-			if vb == va {
-				set = append(set, labels.Label{
-					Name:  ka,
-					Value: va,
-				})
-			}
-		}
-	}
-	sort.Sort(set)
-	return set
-}
-
-// subtract labels set b from labels set a
-func subtract(a, b labels.Labels) labels.Labels {
-
-	set := labels.Labels{}
-	ma := a.Map()
-	mb := b.Map()
-
-	for ka, va := range ma {
-		if vb, ok := mb[ka]; ok {
-			if vb == va {
-				continue
-			}
-		}
-		set = append(set, labels.Label{
-			Name:  ka,
-			Value: va,
-		})
-	}
-	sort.Sort(set)
-	return set
 }
